Added packages/, Erlang 18.3 and Elixir 1.2
packages/ directory will store custom compiled sources that
needs to be deployed on servers are not the property of PreVeil.

In the future we'll have a real debian repository when we have
a real infrastructure but for now we can just download the
package and install it with dpkg -i.

Also I added erlang v18.3 and Elixir 1.2