Add backup_kafka custom tool
Code is located at core:/infrastructure/tools/backup/backup_kafka/

It is to be used with restore_cs_kafka.

backup_kafka will connect to kafka, start a consumer on each partition then writes
all data from the beginning of time to its own partition file inside -backup-dir directory.

It should be noted that backup_kafka will truncate the file when opening it. So it is safe to execute
backup_kafka multiple times.

Usage of ./backup_kafka:
  -backup-dir string
        Backup directory (default "partitions")
  -broker-list string
        Broker list to connect to (default "localhost:9092")
  -topic string
        Topic name
  -version
        Display version and exit